skip file generation for containers that have only list, leaf-list or container children
So far the generator expected that dir_functions will be initialized
and that elements will be added to it when a container child leaf node
is parsed. However for containers that only have list, leaf-list or
container children this would never be the case.
This would result in dirs having an entry for the container path,
without a matching dir_functions entry, which would then cause failures
when files were generated.

Signed-off-by: Juraj Vijtiuk <juraj.vijtiuk@sartura.hr>